+++ Unity.Physics.Hybrid/EntitiesBaking/BakingSystems/ColliderBakingSystem.cs
@@ -174,9 +174,7 @@ namespace Unity.Physics.Authoring
                     });
 
                     AddComponent<PhysicsRootBaked>(entity);
-
                     AddComponent<PhysicsCollider>(entity);
-                    AddBuffer<PhysicsColliderKeyEntityPair>(entity);
 
                     PostProcessTransform(bodyTransform);
                 }