+++ Unity.Physics/Dynamics/Solver/Solver.cs
@@ -297,84 +297,92 @@ namespace Unity.Physics
                 return returnHandles;
             }
 
-            JobHandle handle;
+            JobHandle handle = inputDeps;
 
-            int numPhases = solverSchedulerInfo.NumPhases;
-
-            // Use persistent allocator to allow these to live until the start of next step
+            // early out if there are no work items to process
+            int numPhases = solverSchedulerInfo.NumActivePhases[0];
+            if (numPhases > 0)
             {
-                NativeArray<int> workItemList = solverSchedulerInfo.NumWorkItems;
-
-                //TODO: Change this to Allocator.TempJob when https://github.com/Unity-Technologies/Unity.Physics/issues/7 is resolved
-                JobHandle collisionEventStreamHandle = NativeStream.ScheduleConstruct(out collisionEvents, workItemList, inputDeps, Allocator.Persistent);
-                JobHandle triggerEventStreamHandle = NativeStream.ScheduleConstruct(out triggerEvents, workItemList, inputDeps, Allocator.Persistent);
-                JobHandle impulseEventStreamHandle = NativeStream.ScheduleConstruct(out impulseEvents, workItemList, inputDeps, Allocator.Persistent);
-
-                handle = JobHandle.CombineDependencies(collisionEventStreamHandle, triggerEventStreamHandle, impulseEventStreamHandle);
-
-                float invNumIterations = math.rcp(numIterations);
-
-                var phaseInfoPtrs = (DispatchPairSequencer.SolverSchedulerInfo.SolvePhaseInfo*)NativeArrayUnsafeUtility.GetUnsafeBufferPointerWithoutChecks(solverSchedulerInfo.PhaseInfo);
-
-                float3 gravityNormalized = float3.zero;
-                if (solverStabilizationData.StabilizationHeuristicSettings.EnableSolverStabilization)
+                // Use persistent allocator to allow these to live until the start of next step
                 {
-                    gravityNormalized = math.normalizesafe(solverStabilizationData.Gravity);
-                }
+                    NativeArray<int> workItemList = solverSchedulerInfo.NumWorkItems;
 
-                for (int solverIterationId = 0; solverIterationId < numIterations; solverIterationId++)
-                {
-                    bool firstIteration = solverIterationId == 0;
-                    bool lastIteration = solverIterationId == numIterations - 1;
-                    for (int phaseId = 0; phaseId < numPhases; phaseId++)
-                    {
-                        var job = new ParallelSolverJob
-                        {
-                            JacobiansReader = jacobians.AsReader(),
-                            PhaseIndex = phaseId,
-                            Phases = solverSchedulerInfo.PhaseInfo,
-                            MotionVelocities = dynamicsWorld.MotionVelocities,
-                            SolverStabilizationData = solverStabilizationData,
-                            StepInput = new StepInput
-                            {
-                                InvNumSolverIterations = invNumIterations,
-                                IsFirstIteration = firstIteration,
-                                IsLastIteration = lastIteration,
-                                Timestep = timestep,
-                                InvTimestep = timestep > 0.0f ? 1.0f / timestep : 0.0f
-                            }
-                        };
+                    //TODO: Change this to Allocator.TempJob when https://github.com/Unity-Technologies/Unity.Physics/issues/7 is resolved
+                    JobHandle collisionEventStreamHandle = NativeStream.ScheduleConstruct(out collisionEvents,
+                        workItemList, inputDeps, Allocator.Persistent);
+                    JobHandle triggerEventStreamHandle = NativeStream.ScheduleConstruct(out triggerEvents, workItemList,
+                        inputDeps, Allocator.Persistent);
+                    JobHandle impulseEventStreamHandle = NativeStream.ScheduleConstruct(out impulseEvents, workItemList,
+                        inputDeps, Allocator.Persistent);
 
-                        // Only initialize event writers for last solver iteration jobs
-                        if (lastIteration)
-                        {
-                            job.CollisionEventsWriter = collisionEvents.AsWriter();
-                            job.TriggerEventsWriter = triggerEvents.AsWriter();
-                            job.ImpulseEventsWriter = impulseEvents.AsWriter();
-                        }
+                    handle = JobHandle.CombineDependencies(collisionEventStreamHandle, triggerEventStreamHandle,
+                        impulseEventStreamHandle);
 
-                        // NOTE: The last phase must be executed on a single job since it
-                        // int.MaxValue can't be used as batchSize since 19.1 overflows in that case...
-                        bool isLastPhase = phaseId == numPhases - 1;
-                        int batchSize = isLastPhase ? (int.MaxValue / 2) : 1;
+                    float invNumIterations = math.rcp(numIterations);
 
-                        int* numWorkItems = &(phaseInfoPtrs[phaseId].NumWorkItems);
-                        handle = job.Schedule(numWorkItems, batchSize, handle);
-                    }
+                    var phaseInfoPtrs =
+                        (DispatchPairSequencer.SolverSchedulerInfo.SolvePhaseInfo*)NativeArrayUnsafeUtility
+                            .GetUnsafeBufferPointerWithoutChecks(solverSchedulerInfo.PhaseInfo);
 
-                    // Stabilize velocities
+                    float3 gravityNormalized = float3.zero;
                     if (solverStabilizationData.StabilizationHeuristicSettings.EnableSolverStabilization)
                     {
-                        var stabilizeVelocitiesJob = new StabilizeVelocitiesJob
-                        {
-                            MotionVelocities = dynamicsWorld.MotionVelocities,
-                            SolverStabilizationData = solverStabilizationData,
-                            GravityPerStep = solverStabilizationData.Gravity * timestep,
-                            GravityNormalized = gravityNormalized,
-                            IsFirstIteration = firstIteration
-                        };
+                        gravityNormalized = math.normalizesafe(solverStabilizationData.Gravity);
+                    }
 
-                        handle = stabilizeVelocitiesJob.Schedule(dynamicsWorld.NumMotions, 64, handle);
+                    for (int solverIterationId = 0; solverIterationId < numIterations; solverIterationId++)
+                    {
+                        bool firstIteration = solverIterationId == 0;
+                        bool lastIteration = solverIterationId == numIterations - 1;
+                        for (int phaseId = 0; phaseId < numPhases; phaseId++)
+                        {
+                            var job = new ParallelSolverJob
+                            {
+                                JacobiansReader = jacobians.AsReader(),
+                                PhaseIndex = phaseId,
+                                Phases = solverSchedulerInfo.PhaseInfo,
+                                MotionVelocities = dynamicsWorld.MotionVelocities,
+                                SolverStabilizationData = solverStabilizationData,
+                                StepInput = new StepInput
+                                {
+                                    InvNumSolverIterations = invNumIterations,
+                                    IsFirstIteration = firstIteration,
+                                    IsLastIteration = lastIteration,
+                                    Timestep = timestep,
+                                    InvTimestep = timestep > 0.0f ? 1.0f / timestep : 0.0f
+                                }
+                            };
+
+                            // Only initialize event writers for last solver iteration jobs
+                            if (lastIteration)
+                            {
+                                job.CollisionEventsWriter = collisionEvents.AsWriter();
+                                job.TriggerEventsWriter = triggerEvents.AsWriter();
+                                job.ImpulseEventsWriter = impulseEvents.AsWriter();
+                            }
+
+                            var info = phaseInfoPtrs[phaseId];
+                            // Note: If we have duplicate body indices across batches in this phase we need to process the phase
+                            // sequentially to prevent data races. In this case, we choose a large batch size (batch equal to number of work items)
+                            // to prevent any parallelization of the work.
+                            int batchSize = info.ContainsDuplicateIndices ? info.NumWorkItems : 1;
+                            handle = job.Schedule(info.NumWorkItems, batchSize, handle);
+                        }
+
+                        // Stabilize velocities
+                        if (solverStabilizationData.StabilizationHeuristicSettings.EnableSolverStabilization)
+                        {
+                            var stabilizeVelocitiesJob = new StabilizeVelocitiesJob
+                            {
+                                MotionVelocities = dynamicsWorld.MotionVelocities,
+                                SolverStabilizationData = solverStabilizationData,
+                                GravityPerStep = solverStabilizationData.Gravity * timestep,
+                                GravityNormalized = gravityNormalized,
+                                IsFirstIteration = firstIteration
+                            };
+
+                            handle = stabilizeVelocitiesJob.Schedule(dynamicsWorld.NumMotions, 64, handle);
+                        }
                     }
                 }
             }
@@ -482,7 +490,7 @@ namespace Unity.Physics
 
         [BurstCompile]
         [NoAlias]
-        private struct ParallelSolverJob : IJobParallelForDefer
+        private struct ParallelSolverJob : IJobParallelFor
         {
             [NativeDisableParallelForRestriction]
             public NativeArray<MotionVelocity> MotionVelocities;